task-5: refactoring - placed products-service in separate folder
