+++ products-service/data-sample/stocks.json
