+++ products-service/package.json
@@ -7,10 +7,10 @@
   },
   "homepage": "https://github.com/OnoprienkoArtem/serverless-be#readme",
   "scripts": {
-    "test": "jest",
+    "test": "../node_modules/.bin/jest",
     "deploy": "sls deploy",
-    "load-sample:products": "aws dynamodb batch-write-item --request-items file://products.json --region=eu-west-1",
-    "load-sample:stocks": "aws dynamodb batch-write-item --request-items file://stocks.json --region=eu-west-1"
+    "load-sample:products": "aws dynamodb batch-write-item --request-items file://data-sample/products.json --region=eu-west-1",
+    "load-sample:stocks": "aws dynamodb batch-write-item --request-items file://data-sample/stocks.json --region=eu-west-1"
   },
   "devDependencies": {
     "@babel/core": "^7.21.0",
