+++ products-service/src/__tests__/getProductById.test.js
@@ -4,8 +4,8 @@
 // tests for getProductById
 // Generated by serverless-jest-plugin
 
-const mod = require('./../product-by-id/product-by-id');
-const data = require('../data.json');
+const mod = require('../product-by-id/product-by-id');
+const data = require('../../data-sample/data.json');
 const jestPlugin = require('serverless-jest-plugin');
 const lambdaWrapper = jestPlugin.lambdaWrapper;
 const wrapped = lambdaWrapper.wrap(mod, { handler: 'getProductById' } );
