+++ products-service/src/__tests__/getProducts.test.js
@@ -3,8 +3,8 @@
 // tests for getProducts
 // Generated by serverless-jest-plugin
 
-const data = require('../data.json');
-const mod = require('./../products/products');
+const data = require('../../data-sample/data.json');
+const mod = require('../products/products');
 const jestPlugin = require('serverless-jest-plugin');
 
 const lambdaWrapper = jestPlugin.lambdaWrapper;
