+++ products-service/src/products/products.js
@@ -1,7 +1,7 @@
 // import data from '../data.json';
-import {ErrorResponse} from "../utils/error-handler";
+import {ErrorResponse} from "../../../utils/error-handler";
 import AWS from 'aws-sdk';
-import {mergeProductsAndStocks} from "../utils/utils";
+import {mergeProductsAndStocks} from "../../../utils/utils";
 
 const dynamodb = new AWS.DynamoDB.DocumentClient();
