+++ products-service/webpack.config.js
@@ -4,8 +4,6 @@ const nodeExternals = require("webpack-node-externals");
 module.exports = {
     entry: slsw.lib.entries,
     devtool: "source-map",
-    // Since 'aws-sdk' is not compatible with webpack,
-    // we exclude all node dependencies
     externalsPresets: { node: true },
     externals: [nodeExternals()],
     mode: slsw.lib.webpack.isLocal ? "development" : "production",
@@ -13,16 +11,10 @@ module.exports = {
         minimize: false
     },
     performance: {
-        // Turn off size warnings for entry points
         hints: false
     },
-    // Run babel on all .js files - skip those in node_modules
     module: {
         rules: [
-            // {
-            //     test: /\.json$/,
-            //     loader: 'json-loader'
-            // },
             {
                 test: /\.js$/,
                 loader: "babel-loader",
@@ -31,5 +23,4 @@ module.exports = {
             }
         ],
     },
-    plugins: []
 };
